Ajout du bouton de tri
