+++ index.php
@@ -1,17 +1,7 @@
 <?php
 session_start();
 if (!isset($_SESSION['username'])) {
-    galerieNonConnecte();
-} else {
-    galerieConnecte();
-}
-
-function galerieNonConnecte()
-{
     require_once "pages/galerie.php";
-}
-
-function galerieConnecte()
-{
-    require_once "pages/galerieConn.php";
+} else {
+    require_once "pages/galerie_connecte.php";
 }
